@@ -7,8 +7,8 @@ group 'org.e2immu'
 version '0.0.1'
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_13
-    targetCompatibility = JavaVersion.VERSION_13
+    sourceCompatibility = JavaVersion.VERSION_15
+    targetCompatibility = JavaVersion.VERSION_15
 }
 
 repositories {
@@ -22,6 +22,10 @@ dependencies {
     testImplementation group: 'junit', name: 'junit', version: '4.12'
     testImplementation group: 'io.vertx', name: 'vertx-unit', version: '3.8.5'
     testImplementation group: 'io.vertx', name: 'vertx-web-client', version: '3.8.5'
+}
+
+compileJava {
+    options.compilerArgs.add("--enable-preview")
 }
 
 application {
