@@ -17,29 +17,46 @@
 
 package org.e2immu.analyser.config;
 
+import org.e2immu.analyser.analyser.StatementAnalyser;
+import org.e2immu.analyser.pattern.PatternMatcher;
 import org.e2immu.annotation.Container;
 import org.e2immu.annotation.E2Container;
+
+import java.util.Objects;
+import java.util.function.Supplier;
 
 @E2Container
 public class AnalyserConfiguration {
 
-    public boolean skipTransformations;
+    public final boolean skipTransformations;
+    private final Supplier<PatternMatcher<StatementAnalyser>> patternMatcherSupplier;
 
-    public AnalyserConfiguration(boolean skipTransformations) {
+    public AnalyserConfiguration(boolean skipTransformations, Supplier<PatternMatcher<StatementAnalyser>> patternMatcherSupplier) {
         this.skipTransformations = skipTransformations;
+        this.patternMatcherSupplier = Objects.requireNonNull(patternMatcherSupplier);
+    }
+
+    public PatternMatcher<StatementAnalyser> newPatternMatcher() {
+        return patternMatcherSupplier.get();
     }
 
     @Container(builds = AnalyserConfiguration.class)
     public static class Builder {
         private boolean skipTransformations;
+        private Supplier<PatternMatcher<StatementAnalyser>> patternMatcherSupplier;
 
         public Builder setSkipTransformations(boolean skipTransformations) {
             this.skipTransformations = skipTransformations;
             return this;
         }
 
+        public void setPatternMatcherSupplier(Supplier<PatternMatcher<StatementAnalyser>> patternMatcherSupplier) {
+            this.patternMatcherSupplier = patternMatcherSupplier;
+        }
+
         public AnalyserConfiguration build() {
-            return new AnalyserConfiguration(skipTransformations);
+            return new AnalyserConfiguration(skipTransformations, patternMatcherSupplier == null ?
+                    () -> PatternMatcher.NO_PATTERN_MATCHER : patternMatcherSupplier);
         }
     }
 }
